implimanting meny parts of the user interface style

+++ users/delete_account.php
@@ -2,7 +2,7 @@
 if(isset($_POST['delete_account']))
 
 ?>
-<form action="" method="post">
+<form action="" method="post" class="delete-box">
 
     <input type="submit" value="delete account" class="btn" name="delete">
     <input type="submit" value="not delete account" class="btn" name="not_delete">
@@ -12,24 +12,26 @@ if(isset($_POST['delete_account']))
 <?php 
 
 
+if(isset($_SESSION['username'])){
 
-
-$username_session = $_SESSION['username'];
-if(isset($_POST['delete'])){
-    $delete_query = "delete from `user` where username='$username_session'";
-    $result = mysqli_query($con,$delete_query);
-    if($result){
-   session_unset();
-session_destroy();
-    echo "<Script>alert('account deleted')</Script>";
-    echo "<Script>window.open('../index.php','_Self')</Script>";
+    
+    $username_session = $_SESSION['username'];
+    if(isset($_POST['delete'])){
+        $delete_query = "delete from `user` where username='$username_session'";
+        $result = mysqli_query($con,$delete_query);
+        if($result){
+            session_unset();
+            session_destroy();
+            echo "<Script>alert('account deleted')</Script>";
+            echo "<Script>window.open('../index.php','_Self')</Script>";
     }
 
-
+    
 }
 if(isset($_POST['not_delete'])){
     echo "<Script>window.open('profile.php','_Self')</Script>";
-
+    
+}
 }
